+++ tests/tcp_e2e.rs
@@ -320,8 +320,8 @@ mode = "{tcp_mode}"
         .arg(&config_path)
         .arg("--routes")
         .arg(&routes_path)
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
         .spawn()
         .unwrap();